+++ web/src/lib/components/faces-page/people-search.svelte
